full solution

+++ Codechef/chef_CHEFSTON.cpp
@@ -8,22 +8,21 @@ int main()
     cin >> t;
     while (t--)
     {
-        int n, k;
+        long long n, k;
         cin >> n >> k;
-        int a[n], b[n];
+        long long a[n], b[n];
         for (int i = 0; i < n; i++)
             cin >> a[i];
         for (int i = 0; i < n; i++)
             cin >> b[i];
 
-        int profit = 0, max = 0;
+        long long profit = 0, max = 0;
 
         for (int i = 0; i < n; i++)
         {
             int mul = 1;
             while ((mul * a[i]) <= k)
             {
-
                 mul++;
             }
 
@@ -37,4 +36,5 @@ int main()
     }
     return 0;
 }
+
 //url: https://www.codechef.com/problems/CHEFSTON
